requirements.txt
python-dotenv
google-generativeai
elevenlabs
vertexai
pydantic